1._file_types/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Marianna <mana998@azet.sk>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.81.0"
uvicorn = "^0.18.3"
pandas = "^1.4.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
